pyproject.toml
[project]
name = "wsdbtg"
version = "0.1.0"
description = "Telegram bot playground"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.0",
    "asyncpg>=0.30",
    "fitparse>=1.2",
    "gpxpy>=1.6",
    "lxml>=5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
venvPath = "."
venv = ".venv"
